mining/requirements.txt
pymongo
scrapy
pandas
progressbar2
beautifulsoup4
wptools
msgpack-python
